feat: :sparkles: refactor user use-case paths for improved organization
