+++ src/modules/user/application/use-cases/commands/create-user/create-user.use-case.ts
@@ -1,26 +1,26 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
 
-import { ArchiveUserCommand } from '@user/application/commands/use-cases/archive-user/archive-user.command';
+import { CreateUserCommand } from '@user/application/use-cases/commands/create-user/create-user.command';
 import { UserAggregate } from '@user/domain/aggregates/user.aggregate';
 import { USER_REPOSITORY } from '@user/domain/constants/injection-tokens';
 import { IUserRepositoryContract } from '@user/domain/contracts/user-repository.contract';
-import { ArchiveUserDomainService } from '@user/domain/domain-services/archive-user.domain-service';
+import { CreateUserDomainService } from '@user/domain/domain-services/create-user.domain-service';
 
 @Injectable()
-@CommandHandler(ArchiveUserCommand)
-export class ArchiveUserUseCase implements ICommandHandler<ArchiveUserCommand> {
+@CommandHandler(CreateUserCommand)
+export class CreateUserUseCase implements ICommandHandler<CreateUserCommand> {
   constructor(
-    private readonly archiveUserDomainService: ArchiveUserDomainService,
+    private readonly createUserDomainService: CreateUserDomainService,
     private readonly publisher: EventPublisher,
     @Inject(USER_REPOSITORY)
     private readonly repository: IUserRepositoryContract,
   ) {}
 
-  async execute(command: ArchiveUserCommand): Promise<UserAggregate> {
-    const { uuid } = command;
+  async execute(command: CreateUserCommand): Promise<UserAggregate> {
+    const { user, accounts, profile } = command;
 
-    const userAggregate = await this.archiveUserDomainService.go(uuid);
+    const userAggregate = await this.createUserDomainService.go(user, accounts, profile);
     const userContext = this.publisher.mergeObjectContext(userAggregate);
 
     await this.repository.persist(userAggregate);
