+++ src/modules/user/application/use-cases/commands/destroy-user/destroy-user.command.ts
