+++ src/modules/user/application/use-cases/commands/destroy-user/destroy-user.use-case.ts
@@ -1,29 +1,29 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
 
-import { CreateUserCommand } from '@user/application/commands/use-cases/create-user/create-user.command';
+import { DestroyUserCommand } from '@user/application/use-cases/commands/destroy-user/destroy-user.command';
 import { UserAggregate } from '@user/domain/aggregates/user.aggregate';
 import { USER_REPOSITORY } from '@user/domain/constants/injection-tokens';
 import { IUserRepositoryContract } from '@user/domain/contracts/user-repository.contract';
-import { CreateUserDomainService } from '@user/domain/domain-services/create-user.domain-service';
+import { DestroyUserDomainService } from '@user/domain/domain-services/destroy-user.domain-service';
 
 @Injectable()
-@CommandHandler(CreateUserCommand)
-export class CreateUserUseCase implements ICommandHandler<CreateUserCommand> {
+@CommandHandler(DestroyUserCommand)
+export class DestroyUserUseCase implements ICommandHandler<DestroyUserCommand> {
   constructor(
-    private readonly createUserDomainService: CreateUserDomainService,
+    private readonly DestroyUserDomainService: DestroyUserDomainService,
     private readonly publisher: EventPublisher,
     @Inject(USER_REPOSITORY)
     private readonly repository: IUserRepositoryContract,
   ) {}
 
-  async execute(command: CreateUserCommand): Promise<UserAggregate> {
-    const { user, accounts, profile } = command;
+  async execute(command: DestroyUserCommand): Promise<UserAggregate> {
+    const { uuid } = command;
 
-    const userAggregate = await this.createUserDomainService.go(user, accounts, profile);
+    const userAggregate = await this.DestroyUserDomainService.go(uuid);
     const userContext = this.publisher.mergeObjectContext(userAggregate);
 
-    await this.repository.persist(userAggregate);
+    await this.repository.destroy(userAggregate);
 
     userContext.commit();
