+++ src/modules/user/application/use-cases/commands/update-user/update-user.use-case.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
 
-import { UpdateUserCommand } from '@user/application/commands/use-cases/update-user/update-user.command';
+import { UpdateUserCommand } from '@user/application/use-cases/commands/update-user/update-user.command';
 import { UserAggregate } from '@user/domain/aggregates/user.aggregate';
 import { USER_REPOSITORY } from '@user/domain/constants/injection-tokens';
 import { IUserRepositoryContract } from '@user/domain/contracts/user-repository.contract';
