+++ src/modules/user/application/use-cases/queries/get-user/get-user.query.ts
